add(Polyfill): polyfill for setPrototypeOf under ie11
