+++ src/index.js
@@ -1,5 +1,6 @@
 import '../assets/css/maptalks-ui.css';
 import '../assets/css/maptalks-control.css';
+import './core/polyfill/setPrototypeOf';
 import { version } from '../package.json';
 export * from './core/Constants';
 export { default as Browser } from './core/Browser';
